% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sahieAPI.R
\name{sahie}
\alias{sahie}
\alias{sahie_county}
\alias{sahie_state}
\alias{sahie_us}
\title{Get SAHIE data}
\usage{
sahie(geo, year = 2014, var = c("NAME", "NUI_PT"), api_key = NULL)

sahie_us(...)

sahie_state(...)

sahie_county(...)
}
\arguments{
\item{geo}{(character) One of "us", "state", and "county".}

\item{year}{(numeric) A four-digits value or vector of values. Possible values: 2006--2014.}

\item{var}{(character) A string or vector of strings. Possible fields: \url{http://api.census.gov/data/timeseries/healthins/sahie/variables.html}}

\item{api_key}{Use \code{sahieAPI::set_api_key("<Your api here>")} to set api_key.}

\item{...}{Pass parameter \code{year} or \code{var} for function \code{sahie_us()}, \code{sahie_state()}, and \code{sahie_county()}. Also see Example section.}
}
\value{
Return a tibble, i.e., modern data frame.
If multiple years are specified, it gives a panel dataset.
}
\description{
Get SAHIE data at various geographic levels.
}
\examples{
\dontrun{
sahieAPI::set_api_key("<Your api here>")

# Get national level uninsurance rate for various income groups in 2014
sahieAPI::sahie(geo = "us", year = 2014, var = c("NAME", "NUI_PT", "IPR_DESC", "IPRCAT", "PCTUI_PT"))

# Or
sahieAPI::sahie_us(year = 2014, var = c("NAME", "NUI_PT", "IPR_DESC", "IPRCAT", "PCTUI_PT"))

# Get 2013--2014 data for natinal level, state-level, and county-level
sahieAPI::sahie_us(year = 2013:2014, var = c("NAME", "NUI_PT", "IPR_DESC", "IPRCAT", "PCTUI_PT"))
sahieAPI::sahie_state(year = 2013:2014, var = c("NAME", "NUI_PT", "IPR_DESC", "IPRCAT", "PCTUI_PT"))
sahieAPI::sahie_county(year = 2013:2014, var = c("NAME", "NUI_PT", "IPR_DESC", "IPRCAT", "PCTUI_PT"))
}
}

